requirements.txt
opencv-python
numpy
tensorflow
mediapipe
spotipy
python-dotenv